requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
annotated-types==0.7.0
anyio==4.4.0
beautifulsoup4==4.12.3
fastapi==0.111.0
fastapi-cli==0.0.4
fastjsonschema==2.20.0
highdicom==0.22.0
httpcore==1.0.5
httptools==0.6.1
importlib_metadata==8.0.0
ipython==8.12.3
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
# keras==3.4.1
Markdown==3.6
numpy==1.26.4
pillow==10.4.0
pillow-jpls==1.3.2
pydantic==2.8.2
pydantic_core==2.20.1
pydicom==2.4.4
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
# tensorflow==2.17.0
uvicorn==0.30.1
uvloop==0.19.0
websockets==12.0